// asset-management/src/generated/models/Cr240_assetsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cr240_assetsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cr240_assetsstatecode = keyof typeof Cr240_assetsstatecode;
export const Cr240_assetsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cr240_assetsstatuscode = keyof typeof Cr240_assetsstatuscode;

export interface Cr240_assetsBase {
  "cr240_AssetCategoryID@odata.bind"?: string;
  cr240_assetid: string;
  cr240_assetname?: string;
  "cr240_LocationID@odata.bind"?: string;
  cr240_purchasedate?: string;
  cr240_value?: number;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Cr240_assetsstatecode;
  statuscode?: Cr240_assetsstatuscode;
  timezoneruleversionnumber?: number;
  "TransactionCurrencyId@odata.bind"?: string;
  utcconversiontimezonecode?: number;
}

export interface Cr240_assets extends Cr240_assetsBase {
  cr240_assetcategoryidname?: string;
  cr240_locationidname?: string;
  cr240_value_base?: number;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  exchangerate?: number;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  transactioncurrencyidname?: string;
  versionnumber?: number;
  cr240_assetcategoryid?: object;
  _cr240_assetcategoryid_value?: string;
  cr240_locationid?: object;
  _cr240_locationid_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
  transactioncurrencyid?: object;
  _transactioncurrencyid_value?: string;
}